+++ src/dynamics/implicit.jl
@@ -10,33 +10,37 @@ implicit_correction!(::DiagnosticVariables, ::PrognosticVariables, ::NoImplicit)
 # SHALLOW WATER MODEL
 export ImplicitShallowWater
 
-"""
-Struct that holds various precomputed arrays for the semi-implicit correction to
-prevent gravity waves from amplifying in the shallow water model.
+"""Struct that holds various precomputed arrays for the semi-implicit correction to
+prevent gravity waves from amplifying in the shallow water model. The implicit time step
+between i-1 and i+1 is controlled by the parameter `α` in the range [0, 1]
+    
+    α = 0   means the gravity wave terms are evaluated at i-1 (forward)
+    α = 0.5 evaluates at i+1 and i-1 (centered implicit)
+    α = 1   evaluates at i+1 (backward implicit)
+    α ∈ [0.5, 1] are also possible which controls the strength of the gravity wave dampening.
+    α = 0.5 slows gravity waves and prevents them from amplifying
+    α > 0.5 will dampen the gravity waves within days to a few timesteps (α=1)
+
+Fields are
 $(TYPEDFIELDS)"""
-@kwdef struct ImplicitShallowWater{NF, VectorType} <: AbstractImplicit
-
-    # DIMENSIONS
-    trunc::Int
-
+@kwdef struct ImplicitShallowWater{NF} <: AbstractImplicit
     "[OPTION] coefficient for semi-implicit computations to filter gravity waves, 0.5 <= α <= 1"
     α::NF = 1
 
-    # PRECOMPUTED ARRAYS, to be initiliased with initialize!
-    H::Base.RefValue{NF} = Ref(zero(NF))        # layer_thickness
-    ξH::Base.RefValue{NF} = Ref(zero(NF))       # = 2αΔt*layer_thickness, store in RefValue for mutability
-    g∇²::VectorType = zeros(NF, trunc+2)        # = gravity*eigenvalues
-    ξg∇²::VectorType = zeros(NF, trunc+2)       # = 2αΔt*gravity*eigenvalues
-    S⁻¹::VectorType = zeros(NF, trunc+2)        # = 1 / (1-ξH*ξg∇²), implicit operator
+    "Gravity [m/s²], taken from model.planet at initialize!"
+    gravity::Base.RefValue{NF} = Ref(zero(NF))
+    
+    "Layer thickness [m], taken from model.atmosphere at initialize!"
+    layer_thickness::Base.RefValue{NF} = Ref(zero(NF))
+    
+    "Time step [s], = αdt = 2αΔt (for leapfrog)"
+    time_step::Base.RefValue{NF} = Ref(zero(NF))
 end
 
 """
 $(TYPEDSIGNATURES)
 Generator using the resolution from `spectral_grid`."""
-function ImplicitShallowWater(spectral_grid::SpectralGrid; kwargs...)
-    (; NF, VectorType, trunc) = spectral_grid
-    return ImplicitShallowWater{NF, VectorType}(; trunc, kwargs...)
-end
+ImplicitShallowWater(SG::SpectralGrid; kwargs...) = ImplicitShallowWater{SG.NF}(; kwargs...)
 
 # function barrier to unpack the constants struct for shallow water
 function initialize!(I::ImplicitShallowWater, dt::Real, ::DiagnosticVariables, model::ShallowWater)
@@ -46,36 +50,15 @@ end
 """
 $(TYPEDSIGNATURES)
 Update the implicit terms in `implicit` for the shallow water model as they depend on the time step `dt`."""
-function initialize!(   
+function initialize!(
     implicit::ImplicitShallowWater,
     dt::Real,                   # time step used [s]
     planet::AbstractPlanet,
     atmosphere::AbstractAtmosphere,
 )
-
-    (; α, H, ξH, g∇², ξg∇², S⁻¹) = implicit          # precomputed arrays to be updated
-    (; gravity) = planet                             # gravitational acceleration [m/s²]
-    (; layer_thickness) = atmosphere                 # shallow water layer thickness [m]
-
-    # implicit time step between i-1 and i+1
-    # α = 0   means the gravity wave terms are evaluated at i-1 (forward)
-    # α = 0.5 evaluates at i+1 and i-1 (centered implicit)
-    # α = 1   evaluates at i+1 (backward implicit)
-    # α ∈ [0.5, 1] are also possible which controls the strength of the gravity wave dampening.
-    # α = 0.5 slows gravity waves and prevents them from amplifying
-    # α > 0.5 will dampen the gravity waves within days to a few timesteps (α=1)
-
-    ξ = α*dt                   # new implicit timestep ξ = α*dt = 2αΔt (for leapfrog) from input dt
-    H[] = layer_thickness      # update H the undisturbed layer thickness without mountains
-    ξH[] = ξ*layer_thickness   # update ξ*H with new ξ, in RefValue for mutability
-
-    # loop over degree l of the harmonics (implicit terms are independent of order m)
-    @inbounds for l in eachindex(g∇², ξg∇², S⁻¹)
-        eigenvalue = -l*(l-1)               # =∇², with without 1/radius², 1-based -l*(l+1) → -l*(l-1)
-        g∇²[l] = gravity*eigenvalue         # doesn't actually change with dt
-        ξg∇²[l] = ξ*g∇²[l]                  # update ξg∇² with new ξ
-        S⁻¹[l] = inv(1 - ξH[]*ξg∇²[l])      # update 1/(1-ξ²gH∇²) with new ξ
-    end
+    implicit.gravity[] = planet.gravity                     # gravitational acceleration [m/s²]
+    implicit.layer_thickness[] = atmosphere.layer_thickness # shallow water layer thickness [m], undisturbed, no mountains
+    implicit.time_step[] = implicit.α*dt                    # new implicit timestep ξ = α*dt = 2αΔt (for leapfrog) from input dt
 end
 
 """
@@ -93,35 +76,32 @@ function implicit_correction!(  diagn::DiagnosticVariables,
     pres_old = progn.pres[1]    # pressure/η at t
     pres_new = progn.pres[2]    # pressure/η at t+dt
 
-    (; g∇², ξg∇², S⁻¹) = implicit
-    H = implicit.H[]              # unpack as it's stored in a RefValue for mutation
-    ξH = implicit.ξH[]            # unpack as it's stored in a RefValue for mutation
+    # unpack with [] as stored in a RefValue for mutation during initialization
+    H = implicit.layer_thickness[]      # layer thickness [m], undisturbed, no mountains
+    g = implicit.gravity[]              # gravitational acceleration [m/s²]
+    ξ = implicit.time_step[]            # new implicit timestep ξ = α*dt = 2αΔt (for leapfrog)
+    
+    lmax, mmax = size(div_tend, OneBased, as=Matrix)
 
-    lmax, mmax = size(div_tend, ZeroBased, as=Matrix)
-    lmax -= 1
-
-    @boundscheck length(S⁻¹) == lmax+2 || throw(BoundsError)
-    @boundscheck length(ξg∇²) == lmax+2 || throw(BoundsError)
-    @boundscheck length(g∇²) == lmax+2 || throw(BoundsError)
-
-    for k in eachmatrix(div_tend)
+    @inbounds for k in eachmatrix(div_tend)
         lm = 0
-        for m in 1:mmax+1
-            for l in m:lmax+1
-                lm += 1     # single index lm corresponding to harmonic l, m with a LowerTriangularMatrix
-                
+        for m in 1:mmax
+            for l in m:lmax
+                lm += 1                     # single index lm corresponding to harmonic l, m with a LowerTriangularMatrix
+                ∇² = -l*(l-1)               # eigenvalue, with without 1/radius², 1-based -l*(l+1) → -l*(l-1)
+
                 # calculate the G = N(Vⁱ) + NI(Vⁱ⁻¹ - Vⁱ) term.
                 # Vⁱ is a prognostic variable at time step i
                 # N is the right hand side of ∂V\∂t = N(V)
                 # NI is the part of N that's calculated semi-implicitily: N = NE + NI
-                G_div = div_tend[lm, k] - g∇²[l]*(pres_old[lm] - pres_new[lm])
+                G_div = div_tend[lm, k] - g*∇²*(pres_old[lm] - pres_new[lm])
                 G_η   = pres_tend[lm] - H*(div_old[lm, k] - div_new[lm, k])
 
                 # using the Gs correct the tendencies for semi-implicit time stepping
-                div_tend[lm, k] = S⁻¹[l]*(G_div - ξg∇²[l]*G_η)
-                pres_tend[lm] = G_η - ξH*div_tend[lm, k]
+                S⁻¹ = inv(1 - ξ^2*H*g*∇²)                   # operator to invert
+                div_tend[lm, k] = S⁻¹*(G_div - ξ*g*∇²*G_η)
+                pres_tend[lm] = G_η - ξ*H*div_tend[lm, k]
             end
-            lm += 1     # loop skips last row
         end
     end
 end
@@ -138,7 +118,7 @@ $(TYPEDFIELDS)"""
     MatrixType,
     TensorType,
 } <: AbstractImplicit
-    
+
     # DIMENSIONS
     "spectral resolution"
     trunc::Int
@@ -150,25 +130,25 @@ $(TYPEDFIELDS)"""
     "time-step coefficient: 0=explicit, 0.5=centred implicit, 1=backward implicit"
     α::NF = 1
 
-    # PRECOMPUTED ARRAYS, to be initiliased with initialize!
+    # PRECOMPUTED ARRAYS, to be initiliazed with initialize!
     "vertical temperature profile, obtained from diagn on first time step"
     temp_profile::VectorType = zeros(NF, nlayers)
 
     "time step 2α*Δt packed in RefValue for mutability"
-    ξ::Base.RefValue{NF} = Ref{NF}(0)       
-    
+    ξ::Base.RefValue{NF} = Ref{NF}(0)
+
     "divergence: operator for the geopotential calculation"
-    R::MatrixType = zeros(NF, nlayers, nlayers)     
-    
+    R::MatrixType = zeros(NF, nlayers, nlayers)
+
     "divergence: the -RdTₖ∇² term excl the eigenvalues from ∇² for divergence"
     U::VectorType = zeros(NF, nlayers)
-    
+
     "temperature: operator for the TₖD + κTₖDlnps/Dt term"
     L::MatrixType = zeros(NF, nlayers, nlayers)
 
     "pressure: vertical averaging of the -D̄ term in the log surface pres equation"
     W::VectorType = zeros(NF, nlayers)
-    
+
     "components to construct L, 1/ 2Δσ"
     L0::VectorType = zeros(NF, nlayers)
 
@@ -188,12 +168,12 @@ $(TYPEDFIELDS)"""
     S::MatrixType = zeros(NF, nlayers, nlayers)
 
     "combined inverted operator: S = 1 - ξ²(RL + UW)"
-    S⁻¹::TensorType = zeros(NF, trunc+1, nlayers, nlayers)   
+    S⁻¹::TensorType = zeros(NF, trunc+1, nlayers, nlayers)
 end
 
 """$(TYPEDSIGNATURES)
 Generator using the resolution from SpectralGrid."""
-function ImplicitPrimitiveEquation(spectral_grid::SpectralGrid, kwargs...) 
+function ImplicitPrimitiveEquation(spectral_grid::SpectralGrid, kwargs...)
     (; NF, VectorType, MatrixType, TensorType, trunc, nlayers) = spectral_grid
     return ImplicitPrimitiveEquation{NF, VectorType, MatrixType, TensorType}(;
         trunc, nlayers, kwargs...)
@@ -212,7 +192,7 @@ end
 
 """$(TYPEDSIGNATURES)
 Initialize the implicit terms for the PrimitiveEquation models."""
-function initialize!(   
+function initialize!(
     implicit::ImplicitPrimitiveEquation,
     dt::Real,                                           # the scaled time step radius*dt
     diagn::DiagnosticVariables{NF},
@@ -228,9 +208,9 @@ function initialize!(
     (; Δp_geopot_half, Δp_geopot_full) = geopotential
     (; σ_lnp_A, σ_lnp_B) = adiabatic_conversion
 
-    # use current vertical temperature profile                                     
+    # use current vertical temperature profile
     temp_profile .= diagn.temp_average
-        
+
     # return immediately if temp_profile contains NaRs, model blew up in that case
     all(isfinite.(temp_profile)) || return nothing
 
@@ -249,17 +229,17 @@ function initialize!(
 
     ξ = α*dt                        # dt = 2Δt for leapfrog, but = Δt, Δ/2 in first_timesteps!
     implicit.ξ[] = ξ                # also store in Implicit struct
-    
+
     # DIVERGENCE OPERATORS (called g in Hoskins and Simmons 1975, eq 11 and Appendix 1)
     @inbounds for k in 1:nlayers                # vertical geopotential integration as matrix operator
         R[1:k, k] .= -Δp_geopot_full[k]         # otherwise equivalent to geopotential! with zero orography
-        R[1:k-1, k] .+= -Δp_geopot_half[k]      # incl the minus but excluding the eigenvalues as with U 
+        R[1:k-1, k] .+= -Δp_geopot_half[k]      # incl the minus but excluding the eigenvalues as with U
     end
     U .= -R_dry*temp_profile        # the R_d*Tₖ∇² term excl the eigenvalues from ∇² for divergence
-    
+
     # TEMPERATURE OPERATOR (called τ in Hoskins and Simmons 1975, eq 9 and Appendix 1)
     L0 .= 1 ./ 2σ_levels_thick
-    L2 .= κ*temp_profile.*σ_lnp_A    # factor in front of the div_sum_above term                       
+    L2 .= κ*temp_profile.*σ_lnp_A    # factor in front of the div_sum_above term
     L4 .= κ*temp_profile.*σ_lnp_B    # factor in front of div term in Dlnps/Dt
 
     @inbounds for k in 1:nlayers
@@ -308,18 +288,18 @@ end
 
 """$(TYPEDSIGNATURES)
 Apply the implicit corrections to dampen gravity waves in the primitive equation models."""
-function implicit_correction!(  
+function implicit_correction!(
     diagn::DiagnosticVariables,
     implicit::ImplicitPrimitiveEquation,
     progn::PrognosticVariables,
 )
     # escape immediately if explicit
-    implicit.α == 0 && return nothing   
+    implicit.α == 0 && return nothing
 
     (; nlayers, trunc) = implicit
     (; S⁻¹, R, U, L, W) = implicit
     ξ = implicit.ξ[]
-    
+
     # MOVE THE IMPLICIT TERMS OF THE TEMPERATURE EQUATION FROM TIME STEP i TO i-1
     # geopotential and linear pressure gradient (divergence equation) are already evaluated at i-1
     # so is the -D̄ term for surface pressure in tendencies!
@@ -336,7 +316,7 @@ function implicit_correction!(
             end
         end
     end
-    
+
     # SEMI IMPLICIT CORRECTIONS FOR DIVERGENCE
     # calculate the combined tendency G = G_D + ξRG_T + ξUG_lnps to solve for divergence δD
     (; pres_tend, div_tend) = diagn.tendencies
@@ -351,7 +331,7 @@ function implicit_correction!(
             end
         end
 
-        # 2. the G = G_D + ξRG_T + ξUG_lnps terms using geopot from above 
+        # 2. the G = G_D + ξRG_T + ξUG_lnps terms using geopot from above
         lm = 0
         for m in 1:trunc+1              # loops over all columns/order m
             for l in m:trunc+1          # but skips the lmax+2 degree (1-based)